JdbcUserRepository: string_agg for roles change to ResultExtractor impl, cache test settings fix, mealForm localization h2 fix (edit/add title)

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserRepository.java
@@ -1,10 +1,11 @@
 package ru.javawebinar.topjava.repository.jdbc;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.dao.support.DataAccessUtils;
 import org.springframework.jdbc.core.BatchPreparedStatementSetter;
-import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.ResultSetExtractor;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
@@ -15,27 +16,21 @@ import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.repository.UserRepository;
 
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 
 import static ru.javawebinar.topjava.util.ValidationUtil.checkWithBeanValidation;
 
 @Repository
 @Transactional(readOnly = true)
 public class JdbcUserRepository implements UserRepository {
-
-    private static final BeanPropertyRowMapper<User> ROW_MAPPER = BeanPropertyRowMapper.newInstance(User.class);
-
+    
     private final JdbcTemplate jdbcTemplate;
 
     private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
     private final SimpleJdbcInsert insertUser;
-
-    private static final String INSERT_USER_ROLES_SQL
-            = "INSERT INTO user_roles (ID,EMAIL) VALUES(:id, :email)";
 
     @Autowired
     public JdbcUserRepository(JdbcTemplate jdbcTemplate, NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
@@ -57,16 +52,33 @@ public class JdbcUserRepository implements UserRepository {
         if (user.isNew()) {
             Number newKey = insertUser.executeAndReturnKey(parameterSource);
             user.setId(newKey.intValue());
-        } else if (namedParameterJdbcTemplate.update(
-                "UPDATE users SET name=:name, email=:email, password=:password, " +
-                        "registered=:registered, enabled=:enabled, calories_per_day=:caloriesPerDay WHERE id=:id", parameterSource) == 0) {
-            return null;
+        } else {
+            if (namedParameterJdbcTemplate.update(
+                    "UPDATE users SET name=:name, email=:email, password=:password, " +
+                            "registered=:registered, enabled=:enabled, calories_per_day=:caloriesPerDay WHERE id=:id", parameterSource) == 0) {
+                return null;
+            } else {
+                jdbcTemplate.update("DELETE FROM user_roles WHERE user_id = ?", user.getId());
+            }
         }
 
-        jdbcTemplate.update("DELETE FROM user_roles WHERE user_id = ?", user.getId());
         jdbcTemplate.batchUpdate("INSERT INTO user_roles (user_id, role) VALUES (?, ?)",
-                new UserRolesBatch(user.getId(), user.getRoles()));
+                new BatchPreparedStatementSetter() {
+                    private final List<Role> roles = new ArrayList<>(user.getRoles());
+                    private final Integer userId = user.getId();
 
+                    @Override
+                    public void setValues(PreparedStatement preparedStatement, int i) throws SQLException {
+                        preparedStatement.setInt(1, userId);
+                        preparedStatement.setString(2, roles.get(i).name());
+                    }
+
+                    @Override
+                    public int getBatchSize() {
+                        return roles.size();
+                    }
+                }
+        );
         return user;
     }
 
@@ -78,41 +90,52 @@ public class JdbcUserRepository implements UserRepository {
 
     @Override
     public User get(int id) {
-        List<User> users = jdbcTemplate.query("SELECT u.*, string_agg(ur.role, ',') AS roles FROM users u LEFT JOIN user_roles ur ON ur.user_id = u.id WHERE id=? GROUP BY u.id", ROW_MAPPER, id);
+        List<User> users = jdbcTemplate.query("SELECT * FROM users u LEFT JOIN user_roles ur ON ur.user_id = u.id WHERE id=?", new UserExtractor(), id);
         return DataAccessUtils.singleResult(users);
     }
 
     @Override
     public User getByEmail(String email) {
-//        return jdbcTemplate.queryForObject("SELECT * FROM users WHERE email=?", ROW_MAPPER, email);
-        List<User> users = jdbcTemplate.query("SELECT u.*, string_agg(ur.role, ',') as roles FROM users u LEFT JOIN user_roles ur ON ur.user_id = u.id WHERE email=? GROUP BY u.id", ROW_MAPPER, email);
+        List<User> users = jdbcTemplate.query("SELECT * FROM users u LEFT JOIN user_roles ur ON ur.user_id = u.id WHERE email=?", new UserExtractor(), email);
         return DataAccessUtils.singleResult(users);
     }
 
     @Override
     public List<User> getAll() {
-        return jdbcTemplate.query("SELECT u.*, string_agg(ur.role, ',') as roles FROM users u LEFT JOIN user_roles ur ON ur.user_id = u.id GROUP BY u.id ORDER BY name, email", ROW_MAPPER);
+        return jdbcTemplate.query("SELECT * FROM users u LEFT JOIN user_roles ur ON ur.user_id = u.id ORDER BY name, email", new UserExtractor());
     }
 
-    private static class UserRolesBatch implements BatchPreparedStatementSetter {
-
-        private final int userId;
-        private final List<Role> roles;
-
-        public UserRolesBatch(Integer uid, Set<Role> roles) {
-            this.userId = uid;
-            this.roles = new ArrayList<>(roles);
-        }
-
+    private static class UserExtractor implements ResultSetExtractor<List<User>> {
         @Override
-        public void setValues(PreparedStatement preparedStatement, int i) throws SQLException {
-            preparedStatement.setInt(1, userId);
-            preparedStatement.setString(2, roles.get(i).name());
-        }
+        public List<User> extractData(ResultSet resultSet) throws SQLException, DataAccessException {
+            Map<Integer, User> data = new HashMap<>();
+            while (resultSet.next()) {
+                User user = data.computeIfAbsent(resultSet.getInt("id"), (k) -> {
+                    User newUser = new User();
+                    try {
+                        newUser.setId(resultSet.getInt("id"));
+                        newUser.setName(resultSet.getString("name"));
+                        newUser.setEmail(resultSet.getString("email"));
+                        newUser.setPassword(resultSet.getString("password"));
+                        newUser.setRegistered(resultSet.getDate("registered"));
+                        newUser.setEnabled(resultSet.getBoolean("enabled"));
+                        newUser.setCaloriesPerDay(resultSet.getInt("calories_per_day"));
+                        newUser.setRoles(new HashSet<>());
+                    } catch (SQLException e) {
+                        e.printStackTrace();
+                    }
+                    return newUser;
+                });
 
-        @Override
-        public int getBatchSize() {
-            return roles.size();
+                try {
+                    Set<Role> roles = user.getRoles();
+                    roles.add(Enum.valueOf(Role.class, resultSet.getString("role")));
+                    user.setRoles(roles);
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            return new ArrayList<>(data.values());
         }
     }
 }
